pyproject.toml
[project]
name = "open-swim"
version = "0.1.0"
description = "Open Swim Python project"
readme = "README.md"
authors = [ { name = "Your Name", email = "you@example.com" } ]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyudev>=0.24.3",
]

[project.scripts]
open-swim = "open_swim.main:main"

[tool.uv]
# uv-specific configuration (optional tweaks)
package = true
